feat: modify register ui
